feat(selectors): fix user selectors misgrouped as services
